refactor: 重构聚合 API 支持灵活时间解析和限额控制
- 新增 parse_time() 支持任意时间格式，时间戳改为 u64 防止溢出，新增 aggregate_limit 配置
- 响应格式分离 summary 字段，支持 time_range/start_time/end_time 灵活组合
- 更新文档和 TypeScript 类型，86 个测试全部通过

+++ crates/statistics/src/cli/formatter.rs
@@ -51,7 +51,7 @@ fn format_table(events: &[RoutingEvent]) -> String {
 
     // Rows
     for event in events {
-        let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(event.timestamp)
+        let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(event.timestamp as i64)
             .unwrap_or_default()
             .with_timezone(&Local);
 
@@ -86,7 +86,7 @@ fn format_json(events: &[RoutingEvent]) -> String {
     let json_events: Vec<EventJson> = events
         .iter()
         .map(|e| {
-            let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(e.timestamp)
+            let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(e.timestamp as i64)
                 .unwrap_or_default()
                 .with_timezone(&Local);
 
@@ -119,7 +119,7 @@ fn format_csv(events: &[RoutingEvent]) -> String {
 
     // Rows
     for event in events {
-        let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(event.timestamp)
+        let timestamp: DateTime<Local> = DateTime::from_timestamp_millis(event.timestamp as i64)
             .unwrap_or_default()
             .with_timezone(&Local);
